requirements.txt
beautifulsoup4==4.12.2
google_api_python_client==2.84.0
google_auth_oauthlib==0.5.2
oauth2client==4.1.3
python-dotenv==1.0.0
requests==2.28.2
